requirements.txt
streamlitstreamlit
pandas
matplotlib
scikit-learn
